Move features to own file
