+++ packages/roosterjs-editor-plugins/lib/plugins/ContentEdit/features/markdownFeatures.ts
@@ -1,13 +1,4 @@
-import {
-    cacheGetEventData,
-    createRange,
-    isNodeEmpty,
-    Position,
-    safeInstanceOf,
-    splitBalancedNodeRange,
-    unwrap,
-    wrap,
-} from 'roosterjs-editor-dom';
+import { cacheGetEventData, createRange, Position, wrap } from 'roosterjs-editor-dom';
 import type { CompatibleKeys } from 'roosterjs-editor-types/lib/compatibleTypes';
 import {
     BuildInEditFeature,
@@ -181,73 +172,6 @@ const MarkdownInlineCode: BuildInEditFeature<PluginKeyboardEvent> = generateBasi
     false /* useShiftKey */
 );
 
-const RemoveCodeWhenEnterOnEmptyLine: BuildInEditFeature<PluginKeyboardEvent> = {
-    keys: [Keys.ENTER],
-    shouldHandleEvent: (event, editor) => {
-        const childOfCode = cacheGetCodeChild(event, editor);
-        return childOfCode && isNodeEmpty(childOfCode);
-    },
-    handleEvent: (event, editor) => {
-        event.rawEvent.preventDefault();
-        editor.addUndoSnapshot(
-            () => {
-                splitCode(event, editor);
-            },
-            null /* changeSource */,
-            true /* canUndoByBackspace */
-        );
-    },
-};
-
-const RemoveCodeWhenBackspaceOnEmptyFirstLine: BuildInEditFeature<PluginKeyboardEvent> = {
-    keys: [Keys.BACKSPACE],
-    shouldHandleEvent: (event, editor) => {
-        const childOfCode = cacheGetCodeChild(event, editor);
-        return childOfCode && isNodeEmpty(childOfCode) && !childOfCode.previousSibling;
-    },
-    handleEvent: (event, editor) => {
-        event.rawEvent.preventDefault();
-        editor.addUndoSnapshot(() => splitCode(event, editor));
-    },
-};
-
-function cacheGetCodeChild(event: PluginKeyboardEvent, editor: IEditor): Node | null {
-    return cacheGetEventData(event, 'CODE_CHILD', () => {
-        const codeElement = editor.getElementAtCursor('code');
-        if (codeElement) {
-            const pos = editor.getFocusedPosition();
-            const block = pos && editor.getBlockElementAtNode(pos.normalize().node);
-            if (block) {
-                const node =
-                    block.getStartNode() == codeElement.parentNode
-                        ? block.getStartNode()
-                        : block.collapseToSingleElement();
-                return isNodeEmpty(node) ? node : null;
-            }
-        }
-
-        return null;
-    });
-}
-
-function splitCode(event: PluginKeyboardEvent, editor: IEditor) {
-    const currentContainer = cacheGetCodeChild(event, editor);
-    if (!safeInstanceOf(currentContainer, 'HTMLElement')) {
-        return;
-    }
-    const codeChild = currentContainer.querySelector('code');
-    if (!codeChild) {
-        const codeParent = splitBalancedNodeRange(currentContainer);
-        unwrap(codeParent);
-        const preParent = splitBalancedNodeRange(currentContainer);
-        unwrap(preParent);
-    } else {
-        //Content model
-        unwrap(codeChild);
-    }
-    editor.select(currentContainer, PositionType.Begin);
-}
-
 /**
  * @internal
  */
@@ -259,6 +183,4 @@ export const MarkdownFeatures: Record<
     markdownItalic: MarkdownItalic,
     markdownStrikethru: MarkdownStrikethrough,
     markdownInlineCode: MarkdownInlineCode,
-    removeCodeWhenEnterOnEmptyLine: RemoveCodeWhenEnterOnEmptyLine,
-    removeCodeWhenBackspaceOnEmptyFirstLine: RemoveCodeWhenBackspaceOnEmptyFirstLine,
 };
