+++ packages/roosterjs-editor-types/lib/interface/ContentEditFeatureSettings.ts
@@ -170,18 +170,6 @@ export interface MarkdownFeatureSettings {
      * @default true
      */
     markdownInlineCode: boolean;
-
-    /**
-     * When inside a code block, exit the code block by pressing Enter twice, or once on an empty line
-     * @default true
-     */
-    removeCodeWhenEnterOnEmptyLine: boolean;
-
-    /**
-     * When inside an empty code block (or an empty first line), exit the code block by pressing Backspace
-     * @default true
-     */
-    removeCodeWhenBackspaceOnEmptyFirstLine: boolean;
 }
 
 /**
@@ -281,6 +269,20 @@ export interface TextFeatureSettings {
     autoHyphen: boolean;
 }
 
+export interface CodeFeatureSettings {
+    /**
+     * When inside a code block, exit the code block by pressing Enter twice, or once on an empty line
+     * @default true
+     */
+    removeCodeWhenEnterOnEmptyLine: boolean;
+
+    /**
+     * When inside an empty code block (or an empty first line), exit the code block by pressing Backspace
+     * @default true
+     */
+    removeCodeWhenBackspaceOnEmptyFirstLine: boolean;
+}
+
 /**
  * A list to specify whether each of the listed content edit features is enabled
  */
@@ -294,4 +296,5 @@ export default interface ContentEditFeatureSettings
         CursorFeatureSettings,
         MarkdownFeatureSettings,
         EntityFeatureSettings,
-        TextFeatureSettings {}
+        TextFeatureSettings,
+        CodeFeatureSettings {}
