+++ packages/roosterjs-editor-types/lib/interface/index.ts
@@ -105,6 +105,7 @@ export {
     StructuredNodeFeatureSettings,
     TableFeatureSettings,
     TextFeatureSettings,
+    CodeFeatureSettings,
 } from './ContentEditFeatureSettings';
 export { default as CustomReplacement } from './CustomReplacement';
 export { default as UndoSnapshotsService } from './UndoSnapshotsService';
